{ivhd} Added CUDA to build path. Fix : Project building without any warnings.

+++ ivhd/libivhd/source/embed/cast/ivhd/CasterAdam.cpp
@@ -17,16 +17,17 @@ namespace ivhd::embed::cast::ivhd
 		auto& forces = m_ext_particleSystem.calculationData()->m_force;
 
 		const auto it = m_ext_particleSystem.step();
+
 		for (auto i = 0; i < m_ext_particleSystem.countParticles(); i++)
 		{
-			decGrad[i].x = decGrad[i].x * B2 + (1.0f - powf(B2, it)) * forces[i].x * forces[i].x;
-			decGrad[i].y = decGrad[i].y * B2 + (1.0f - powf(B2, it)) * forces[i].y * forces[i].y;
+			decGrad[i].x = decGrad[i].x * B2 + (1.0f - powf(B2, static_cast<float>(it))) * forces[i].x * forces[i].x;
+			decGrad[i].y = decGrad[i].y * B2 + (1.0f - powf(B2, static_cast<float>(it))) * forces[i].y * forces[i].y;
 
-			decDelta[i].x = decDelta[i].x * B1 + (1.0f - powf(B1, it)) * forces[i].x;
-			decDelta[i].y = decDelta[i].y * B1 + (1.0f - powf(B1, it)) * forces[i].y;
+			decDelta[i].x = decDelta[i].x * B1 + (1.0f - powf(B1, static_cast<float>(it))) * forces[i].x;
+			decDelta[i].y = decDelta[i].y * B1 + (1.0f - powf(B1, static_cast<float>(it))) * forces[i].y;
 
-			const auto deltax = LEARNING_RATE * (decDelta[i].x / (1.0f - B1)) / (EPS + sqrtf(decGrad[i].x / (1.0 - B2)));
-			const auto deltay = LEARNING_RATE * (decDelta[i].y / (1.0f - B1)) / (EPS + sqrtf(decGrad[i].y / (1.0 - B2)));
+			const auto deltax = LEARNING_RATE * (decDelta[i].x / (1.0f - B1)) / (EPS + sqrtf(decGrad[i].x / (1.0f - B2)));
+			const auto deltay = LEARNING_RATE * (decDelta[i].y / (1.0f - B1)) / (EPS + sqrtf(decGrad[i].y / (1.0f - B2)));
 
 			positions[i].x += deltax;
 			positions[i].y += deltay;
